# docs/swagger.yaml
basePath: /
definitions:
  application.CreateSubscriptionCommand:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  domain.Subscription:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  response.MessageResponse:
    properties:
      message:
        example: success message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a high performance notification system API
  title: Notifier API
  version: "1.0"
paths:
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Create a subscription
      parameters:
      - description: Subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/application.CreateSubscriptionCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    get:
      consumes:
      - application/json
      description: Get a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscription'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a subscription by ID
      tags:
      - subscriptions
swagger: "2.0"